refactor: set up framework to support importing outline textures
